challenges/ai/helpful-assistant/requirements.txt
blinker==1.8.2 ; python_version >= "3.12" and python_version < "4.0"
cachetools==5.5.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
cloudevents==1.11.0 ; python_version >= "3.12" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
deprecated==1.2.14 ; python_version >= "3.12" and python_version < "4.0"
deprecation==2.1.0 ; python_version >= "3.12" and python_version < "4"
flask==3.0.3 ; python_version >= "3.12" and python_version < "4.0"
functions-framework==3.8.1 ; python_version >= "3.12" and python_version < "4"
google-api-core==2.19.2 ; python_version >= "3.12" and python_version < "4.0"
google-api-core[grpc]==2.19.2 ; python_version >= "3.12" and python_version < "4.0"
google-auth==2.34.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-appengine-logging==1.4.5 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-audit-log==0.3.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-core==2.4.1 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-logging==3.11.2 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos==1.65.0 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos[grpc]==1.65.0 ; python_version >= "3.12" and python_version < "4.0"
grpc-google-iam-v1==0.13.1 ; python_version >= "3.12" and python_version < "4.0"
grpcio-status==1.66.1 ; python_version >= "3.12" and python_version < "4.0"
grpcio==1.66.1 ; python_version >= "3.12" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.8 ; python_version >= "3.12" and python_version < "4.0"
importlib-metadata==8.4.0 ; python_version >= "3.12" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-api==1.27.0 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.12" and python_version < "4.0"
proto-plus==1.24.0 ; python_version >= "3.12" and python_version < "4.0"
protobuf==5.28.0 ; python_version >= "3.12" and python_version < "4.0"
pyasn1-modules==0.4.0 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4"
urllib3==2.2.2 ; python_version >= "3.12" and python_version < "4.0"
watchdog==5.0.0 ; python_version >= "3.12" and python_version < "4"
werkzeug==3.0.4 ; python_version >= "3.12" and python_version < "4.0"
wrapt==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
zipp==3.20.1 ; python_version >= "3.12" and python_version < "4.0"
